final jing2

+++ cell04/ex02/calculator.py
@@ -1,5 +1,5 @@
-a = float(input("Enter first number: "))
-b = float(input("Enter second number: "))
+a = int(input("Enter first number: "))
+b = int(input("Enter second number: "))
 
 print("Addition:", a + b)
 print("Subtraction: ", a - b)
